lab-grader/requirements.txt
fastapi==0.78.0
aiogoogle==4.2.0
uvicorn==0.18.1
gidgethub==5.2.0
PyYAML==6.0
python-dateutil==2.8.2
slowapi==0.1.5